Move library implementation into v2 subdirectory
That seems to be how go modules work for versions greater than 1.

See this stack overlow answer https://stackoverflow.com/a/53380636.
That's also something hinted in https://blog.golang.org/migrating-to-go-modules but not explained in details.
