+++ README.md
@@ -28,8 +28,8 @@ import (
     "os"
     "strings"
 
-    "github.com/optiopay/kafka"
-    "github.com/optiopay/kafka/proto"
+    "github.com/optiopay/kafka/v2"
+    "github.com/optiopay/kafka/v2/proto"
 )
 
 const (
