+++ kafkatest/broker.go
@@ -6,8 +6,8 @@ import (
 	"sync"
 	"time"
 
-	"github.com/optiopay/kafka"
-	"github.com/optiopay/kafka/proto"
+	"github.com/optiopay/kafka/v2"
+	"github.com/optiopay/kafka/v2/proto"
 )
 
 var (
