+++ kafkatest/example_test.go
@@ -6,8 +6,8 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/optiopay/kafka"
-	"github.com/optiopay/kafka/proto"
+	"github.com/optiopay/kafka/v2"
+	"github.com/optiopay/kafka/v2/proto"
 )
 
 func ExampleBroker_Producer() {
