+++ kafkatest/server.go
@@ -13,7 +13,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/optiopay/kafka/proto"
+	"github.com/optiopay/kafka/v2/proto"
 )
 
 type topicOffset struct {
