+++ v2/distributing_producer.go
@@ -8,7 +8,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/optiopay/kafka/proto"
+	"github.com/optiopay/kafka/v2/proto"
 )
 
 // DistributingProducer is the interface similar to Producer, but never require
