+++ v2/integration/kafka-docker/Dockerfile
