+++ v2/integration/kafka-docker/LICENSE
