+++ v2/integration/kafka-docker/docker-compose-single-broker.yml
