+++ v2/integration/kafka-docker/docker-compose.yml
