+++ v2/proto/snappy_test.go
